fix(build): resolved build error and clean up comments
