+++ src/app/timeline-context.tsx
@@ -7,7 +7,6 @@ import {
   ReactNode,
 } from "react";
 
-// Define the shape of the context state
 interface TimelineContextType {
   graduationDate: string;
   setGraduationDate: (date: string) => void;
@@ -27,25 +26,18 @@ interface TimelineContextType {
   setH1bCapGapEndDate: (date: string) => void;
 }
 
-// Create the context with a default undefined value (it will be set by the provider)
 const TimelineContext = createContext<TimelineContextType | undefined>(
   undefined
 );
 
-// Define the props for the TimelineProvider component
 interface TimelineProviderProps {
   children: ReactNode;
 }
 
-/**
- * TimelineProvider component.
- * Manages all the shared timeline state and provides it to its children.
- * State is now session-based (in-memory only), not persisted to local storage.
- */
 export const TimelineProvider: React.FC<TimelineProviderProps> = ({
   children,
 }) => {
-  // Initialize state with empty strings or 0, no longer loading from local storage
+
   const [graduationDate, setGraduationDate] = useState<string>("");
   const [initialOptStartDate, setInitialOptStartDate] = useState<string>("");
   const [initialOptEndDate, setInitialOptEndDate] = useState<string>("");
@@ -55,10 +47,6 @@ export const TimelineProvider: React.FC<TimelineProviderProps> = ({
   const [h1bPetitionDate, setH1bPetitionDate] = useState<string>("");
   const [h1bCapGapEndDate, setH1bCapGapEndDate] = useState<string>("");
 
-  // Removed the useEffect that saved state to local storage.
-  // Data is now only held in memory for the current session.
-
-  // The value object that will be provided to consumers
   const value = {
     graduationDate,
     setGraduationDate,
@@ -85,10 +73,7 @@ export const TimelineProvider: React.FC<TimelineProviderProps> = ({
   );
 };
 
-/**
- * Custom hook to consume the TimelineContext.
- * Throws an error if used outside of a TimelineProvider.
- */
+
 export const useTimeline = () => {
   const context = useContext(TimelineContext);
   if (context === undefined) {
