+++ src/components/navbar/mobile-nav-items.tsx
@@ -13,15 +13,14 @@ import {
 import { cn } from "@/lib/utils";
 import { Menu } from "lucide-react";
 import Link from "next/link";
-import { usePathname, useSearchParams } from "next/navigation";
-import { proTipExamples } from "../main/examples-data/pro-tips-examples-data";
+import { usePathname } from "next/navigation";
 import { navItems } from "./constants/navbar.constants";
 import ThemeToggle from "./theme-toggle";
+import { ExamplesDropdownContent } from "./examples-dropdown-content";
+import { Suspense } from "react";
 
 const MobileNavItems = () => {
   const pathname = usePathname();
-  const searchParams = useSearchParams();
-  const scenarioId = searchParams.get("scenario");
 
   return (
     <div className="lg:hidden">
@@ -65,57 +64,9 @@ const MobileNavItems = () => {
                             {child.name}
                           </DropdownMenuSubTrigger>
                           <DropdownMenuSubContent>
-                            <DropdownMenuItem
-                              className="font-bold text-foreground cursor-default"
-                              disabled
-                            >
-                              Initial OPT Scenarios
-                            </DropdownMenuItem>
-                            {proTipExamples
-                              .filter((ex) => ex.category === "Initial OPT")
-                              .map((example) => (
-                                <Link
-                                  key={example.id}
-                                  href={`${child.path}?scenario=${example.id}`}
-                                >
-                                  <DropdownMenuItem
-                                    className={cn(
-                                      "cursor-pointer",
-                                      scenarioId === example.id
-                                        ? "bg-accent text-accent-foreground"
-                                        : ""
-                                    )}
-                                  >
-                                    {example.title}
-                                  </DropdownMenuItem>
-                                </Link>
-                              ))}
-                            <DropdownMenuSeparator />
-                            <DropdownMenuItem
-                              className="font-bold text-foreground cursor-default"
-                              disabled
-                            >
-                              STEM OPT Scenarios
-                            </DropdownMenuItem>
-                            {proTipExamples
-                              .filter((ex) => ex.category === "STEM OPT")
-                              .map((example) => (
-                                <Link
-                                  key={example.id}
-                                  href={`${child.path}?scenario=${example.id}`}
-                                >
-                                  <DropdownMenuItem
-                                    className={cn(
-                                      "cursor-pointer",
-                                      scenarioId === example.id
-                                        ? "bg-accent text-accent-foreground"
-                                        : ""
-                                    )}
-                                  >
-                                    {example.title}
-                                  </DropdownMenuItem>
-                                </Link>
-                              ))}
+                            <Suspense fallback={<DropdownMenuItem>Loading...</DropdownMenuItem>}>
+                              <ExamplesDropdownContent childPath={child.path} />
+                            </Suspense>
                           </DropdownMenuSubContent>
                         </DropdownMenuSub>
                       );
